Switched from PyYaml to custom parser.
We can get rid of the bothersome PyYaml dependency by using a very
simple regexp-based parser, which happens to be a lot faster than
a full YAML parser. It also allows us to work around some oddities
in the way beanstalkd emits YAML: if a tube name looks like a
number, the new parser will still parse it as a string, which is
probably what users expect to happen.