Add constant-time and exponential backoff options.
When using a connection with exponential or constant backoff, you
should pass connect_automatically=False to the constructor, and
call connect() manually. This may fail. If you then call connect()
again, there will be a small delay before it attempts to reconnect.
If it fails again, the next delay will be larger, and so on,
exponentially. A successful connection resets this.